feat/Time decayed scoring
Added a scoring function to paper views.  If
a user selects popularity, it will sort the
returned results (best match) by a sort
of the papers interaction over time.
The older the interaction, the more penalized the
time.  This method will focus on popularity
of the most recent papers allowing you more
search flexibility